# backend/tests/contract/openapi.json
{
  "components": {
    "schemas": {
      "AdminSessionRequest": {
        "properties": {
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "AdminSessionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionRequest": {
        "properties": {
          "customer_id": {
            "minLength": 1,
            "title": "Customer Id",
            "type": "string"
          },
          "document_number": {
            "minLength": 1,
            "title": "Document Number",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "document_number"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "session_token": {
            "title": "Session Token",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "session_token",
          "token_type",
          "expires_in"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nMock banking service (tool layer) for the Factored AI & Data Hackathon 2026\nprototype. Sandbox limitations, stated explicitly:\n\n- Login is a test session: `POST /session` checks `customer_id` +\n  `document_number` against `gold.customers` and issues a short-lived JWT\n  (HS256). No real identity provider is behind it.\n- Admin login (`POST /admin/session`) checks credentials against `ADMIN_USERS`\n  (bcrypt hashes, in-memory rate limit, single-process sandbox limitation).\n  Admins work an inbox of escalated cases with audited transitions\n  (claim -> in_progress -> closed); they never set `auto_resolved`.\n- `GET /meta/demo-scenarios` is a public demo aid: it surfaces only the fields\n  the test login itself asks for (customer_id, document_number, name), never\n  fraud columns.\n- Reads come from the read-only `gold.*` tables published by the dbt pipeline;\n  operational tables (`app.*`) are built by the versioned SQL migrations in\n  `app/migrations`, applied at startup under an advisory lock.\n- `/v1` is what the web app uses (user name and password login, products,\n  paginated history, summary). The routes at the root are the agent's.\n- No money is ever moved: disputes only explain, document and escalate.\n- Fraud ground truth (`is_fraud`/`fraud_score`) is not part of `gold.transactions`\n  and is never exposed here.\n",
    "title": "mock-banking-service",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/session": {
      "post": {
        "operationId": "create_admin_session_admin_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Admin Session",
        "tags": [
          "session",
          "admin"
        ]
      }
    },
...
    "/session": {
      "post": {
        "operationId": "create_session_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "session"
        ]
      }
    },
...
  }
}